pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelling"
version = "0.1.0"
description = "Command-line tool for converting images into pixel-style outputs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow",
]

[project.scripts]
pixelling = "pixelling.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
